Include the block variations on the inserter selector
