+++ packages/block-editor/src/components/block-types-list/index.js
@@ -13,7 +13,6 @@ import { useEffect } from '@wordpress/element';
  * Internal dependencies
  */
 import InserterListItem from '../inserter-list-item';
-import { includeVariationsInInserterItems } from '../inserter/utils';
 
 function BlockTypesList( {
 	items = [],
@@ -21,14 +20,9 @@ function BlockTypesList( {
 	onHover = () => {},
 	children,
 	label,
-	limit,
 } ) {
 	const composite = useCompositeState();
-	const normalizedItems = includeVariationsInInserterItems( items, limit );
-	const orderId = normalizedItems.reduce(
-		( acc, item ) => acc + '--' + item.id,
-		''
-	);
+	const orderId = items.reduce( ( acc, item ) => acc + '--' + item.id, '' );
 
 	// This ensures the composite state refreshes when the list order changes.
 	useEffect( () => {
@@ -47,7 +41,7 @@ function BlockTypesList( {
 			className="block-editor-block-types-list"
 			aria-label={ label }
 		>
-			{ normalizedItems.map( ( item ) => {
+			{ items.map( ( item ) => {
 				return (
 					<InserterListItem
 						key={ item.id }
