+++ packages/block-editor/src/components/inserter/block-types-tab.js
@@ -147,7 +147,6 @@ export function BlockTypesTab( {
 							onSelect={ onSelectItem }
 							onHover={ onHover }
 							label={ _x( 'Most used', 'blocks' ) }
-							limit={ MAX_SUGGESTED_ITEMS }
 						/>
 					</InserterPanel>
 				) }
